+++ agenta-backend/agenta_backend/routers/evaluation_router.py
@@ -31,7 +31,8 @@ from agenta_backend.services.evaluation_service import (
     create_new_evaluation,
     create_new_evaluation_scenario,
 )
-from agenta_backend.services.db_manager import engine, query, get_user_object
+from agenta_backend.utills.common import engine
+from agenta_backend.services.db_manager import query, get_user_object
 from agenta_backend.models.db_models import EvaluationDB, EvaluationScenarioDB
 from agenta_backend.config import settings
