+++ agenta-backend/agenta_backend/routers/testset_router.py
@@ -16,8 +16,9 @@ from agenta_backend.models.api.testset_model import (
     TestSetOutputResponse,
 )
 from agenta_backend.config import settings
+from agenta_backend.utills.common import engine
 from agenta_backend.models.db_models import TestSetDB
-from agenta_backend.services.db_manager import engine, query, get_user_object
+from agenta_backend.services.db_manager import query, get_user_object
 
 
 upload_folder = "./path/to/upload/folder"
