+++ agenta-backend/agenta_backend/services/user_service.py
@@ -1,7 +1,6 @@
-from agenta_backend.services.db_manager import engine, UserDB
+from agenta_backend.utills.common import engine
+from agenta_backend.models.db_models import UserDB
 from agenta_backend.models.api.user_models import User, UserUpdate
-from agenta_backend.services.organization_service import get_organization
-
 
 async def create_new_user(payload: User) -> UserDB:
     user_instance = UserDB(
