+++ agenta-backend/agenta_backend/utills/common.py
@@ -1,4 +1,6 @@
 import secrets
+from agenta_backend.models.db_engine import DBEngine
+engine = DBEngine(mode="default").engine()
 
 
 def generate_invitation_token(token_length: int = 16):
